+++ src/main/java/at/fhburgenland/services/PflegerService.java
@@ -6,6 +6,7 @@ import jakarta.validation.ConstraintViolationException;
 
 import java.util.ArrayList;
 import java.util.List;
+import at.fhburgenland.entities.*;
 
 public class PflegerService {
     private static EntityManagerFactory emf = Persistence.createEntityManagerFactory("project");
@@ -14,11 +15,40 @@ public class PflegerService {
         // TODO Menu und Logik für Pfleger
     }
 
-    public static void create(Pfleger p){
+    public static void create(Pfleger p,
+                              List<Integer> tierIds,
+                              List<Integer> gehegeIds,
+                              List<Integer> planIds,
+                              List<Integer> inventarIds,
+                              List<Integer> fuehrungIds)
+    {
         EntityManager em = emf.createEntityManager();
         EntityTransaction et = null;
         try {
-            et = em.getTransaction(); et.begin();
+            et = em.getTransaction();
+            et.begin();
+
+            for (Integer tid : tierIds) {
+                Tier t = em.getReference(Tier.class, tid);
+                p.addTier(t);
+            }
+            for (Integer gid : gehegeIds) {
+                Gehege g = em.getReference(Gehege.class, gid);
+                p.addGehege(g);
+            }
+            for (Integer pid : planIds) {
+                Fuetterungsplan f = em.getReference(Fuetterungsplan.class, pid);
+                p.addFuetterungsplan(f);
+            }
+            for (Integer iid : inventarIds) {
+                Inventar i = em.getReference(Inventar.class, iid);
+                p.addInventar(i);
+            }
+            for (Integer fid : fuehrungIds) {
+                Fuehrung f = em.getReference(Fuehrung.class, fid);
+                p.addFuehrung(f);
+            }
+
             em.persist(p);
             et.commit();
         } catch (Exception e){
@@ -45,7 +75,8 @@ public class PflegerService {
         EntityManager em = emf.createEntityManager();
         EntityTransaction et = null;
         try{
-            et = em.getTransaction(); et.begin();
+            et = em.getTransaction();
+            et.begin();
             Pfleger existing = em.find(Pfleger.class, p.getPflegerId());
             existing.setVorname(p.getVorname());
             existing.setNachname(p.getNachname());
@@ -71,9 +102,18 @@ public class PflegerService {
         EntityManager em = emf.createEntityManager();
         EntityTransaction et = null;
         try{
-            et = em.getTransaction(); et.begin();
+            et = em.getTransaction();
+            et.begin();
             Pfleger p = em.find(Pfleger.class, id);
-            em.remove(p);
+
+            if (p != null) {
+                p.getGepflegteTiere().clear();
+                p.getGereinigteGehege().clear();
+                p.getFuetterungsplaene().clear();
+                p.getInventarListe().clear();
+                p.getFuehrungen().clear();
+                em.remove(p);
+            }
             et.commit();
         } catch (ConstraintViolationException cve) {
             if(et.isActive()){
